Moved Tools class to library package
