+++ src/org/usfirst/frc/team4624/robot/library/Tools.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team4624.robot;
+package org.usfirst.frc.team4624.robot.library;
